rename directory from `perplexity` to `perplexityapikey`
