+++ enricher/secrets/secrets.go
@@ -25,7 +25,7 @@ import (
 	"github.com/google/osv-scalibr/plugin"
 	"github.com/google/osv-scalibr/veles"
 	"github.com/google/osv-scalibr/veles/secrets/gcpsak"
-	perplexityapikey "github.com/google/osv-scalibr/veles/secrets/perplexity"
+	perplexityapikey "github.com/google/osv-scalibr/veles/secrets/perplexityapikey"
 )
 
 const (
