+++ veles/secrets/perplexityapikey/detector.go
